[refactor] #23 - add location querys to single module
