+++ server/api/locations.post.ts
@@ -1,11 +1,7 @@
 import type { DrizzleError } from 'drizzle-orm'
-import { and, eq } from 'drizzle-orm'
-import { customAlphabet } from 'nanoid'
 import slugify from 'slug'
-import db from '~/lib/db'
-import { InsertLocation, location } from '~/lib/db/schema'
-
-const nanoid = customAlphabet('1234567890abcdefghijklmnopqrstwxyz', 5)
+import { findLocationByName, findUniqueSlug, insertLocatioQuery } from '~/lib/db/queries/location-queries'
+import { InsertLocation } from '~/lib/db/schema'
 
 export default defineEventHandler(async (event) => {
   if (!event.context.user) {
@@ -29,12 +25,7 @@ export default defineEventHandler(async (event) => {
     }))
   }
 
-  const existingLocation = !!(await db.query.location.findFirst({
-    where: and(
-      eq(location.name, result.data.name),
-      eq(location.userId, event.context.user.id),
-    ),
-  }))
+  const existingLocation = await findLocationByName(result.data, event.context.user.id)
 
   if (existingLocation) {
     return sendError(event, createError({
@@ -43,29 +34,9 @@ export default defineEventHandler(async (event) => {
     }))
   }
 
-  let slug = slugify(result.data.name)
-
-  let existingSlug = !!(await db.query.location.findFirst({
-    where: eq(location.slug, slug),
-  }))
-  // todo optimize this query
-  while (existingSlug) {
-    const id = nanoid()
-    const idSlug = `${slug}-${id}`
-    existingSlug = !!(await db.query.location.findFirst({
-      where: eq(location.slug, idSlug),
-    }))
-    if (!existingSlug) {
-      slug = idSlug
-    }
-  }
+  const slug = await findUniqueSlug(slugify(result.data.name)) || ''
   try {
-    const [created] = await db.insert(location).values({
-      ...result.data,
-      slug,
-      userId: event.context.user.id,
-    }).returning()
-    return created
+    return insertLocatioQuery(result.data, slug, event.context.user.id)
   }
   catch (e) {
     const error = e as DrizzleError
